refactor: align folder names with file responsibilities
- Move obf/board-import.{ts,test.ts} → storage/ (file persists to IndexedDB; placement now matches its primary job)
- Rename obf/obf-mapper.{ts,test.ts} → obf/mapper.{ts,test.ts} (drop the obf/obf stutter)
- Flatten app/hooks/usePageTitle.ts → app/usePageTitle.ts (remove single-file folder)
- Flatten suggestions/components/ToneSelector.{tsx,test.tsx} → suggestions/ (remove single-file components/ folder)
- Rename shared/testing/fixtures/ → shared/testing/sample-boards/ (avoid file/folder homonym with fixtures.ts)
- Update imports, internal constant FIXTURES_DIR → SAMPLE_BOARDS_DIR, and architecture.md link

No behavior, type, or public-API changes. 159 tests pass.
