+++ src/features/board/suggestions/SuggestionBar.tsx
@@ -2,7 +2,7 @@ import Box from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import Stack from "@mui/material/Stack";
 import type { SuggestionTone } from "../types";
-import { ToneSelector } from "./components/ToneSelector";
+import { ToneSelector } from "./ToneSelector";
 
 export interface SuggestionBarProps {
   suggestions: string[];
